+++ kernel/test/tests.c
@@ -11,7 +11,7 @@ void kernel_tests(void)
 
      // test_vmm();
 
-     test_heap();
+     //test_heap();
 
-    test_summary();
+     //test_summary();
 }